seperate fundamentals and gmsh wrapper
